feat(template): template for Atomic Studio packaging projects

+++ flake.nix
@@ -35,7 +35,8 @@
         ];
       });
 
-      checks = forEachSupportedSystem ({ pkgs }: {
+      checks = forEachSupportedSystem ({ pkgs }: rec {
+        default = pre-commit-check;
         pre-commit-check = nix-pre-commit-hooks.lib.${pkgs.system}.run {
           src = ./.;
 
@@ -47,7 +48,6 @@
             commitizen.enable = true;
             markdownlint.enable = true;
           };
-
           settings = {
             markdownlint.config = {
               "MD013" = {
@@ -73,22 +73,28 @@
 
       packages = forEachSupportedSystem ({ pkgs }: {
         generate-sbkey = pkgs.writers.writeNuBin "sbkey-generator" ''
-          def main [--folder_name (-f): string] {
-            if $folder_name != null {
-              mkdir $folder_name
-            } else {
-              mkdir result
-            }
+          	  # Generate secure boot keys for any kernel signing effort you need
+                    def main [--folder_name (-f): string] {
+                      if $folder_name != null {
+                        mkdir $folder_name
+                      } else {
+                        mkdir result
+                      }
 
-            ${pkgs.lib.getExe pkgs.openssl} req -new -x509 -newkey rsa:2048 -nodes -days 36500 -outform DER -keyout "result/MOK.priv" -out "result/MOK.der"
-          }
+                      ${pkgs.lib.getExe pkgs.openssl} req -new -x509 -newkey rsa:2048 -nodes -days 36500 -outform DER -keyout "result/MOK.priv" -out "result/MOK.der"
+                    }
         '';
 
         cosign-generate = pkgs.writers.writeNuBin "cosign-generate" ''
-          echo "DO NOT add any password, this will break your CI jobs!"
-          ${pkgs.cosign}/bin/cosign generate-key-pair
-          open cosign.key | ${pkgs.lib.getExe pkgs.gh} secret set SIGNING_SECRET --app actions
-          rm cosign.key
+          	  # Generate a private and public cosign key for container signing usage!
+          	  def main [--keep] {
+          	    echo "DO NOT add any password, this will break your CI jobs!"
+                      ${pkgs.cosign}/bin/cosign generate-key-pair
+                      open cosign.key | ${pkgs.lib.getExe pkgs.gh} secret set SIGNING_SECRET --app actions
+          	    if $keep == null {
+          	    rm cosign.key
+          	    }
+          	  }
         '';
       });
